React Router DOM: `routing.kt` components separation
