+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/BrowserRouter.kt
@@ -1,0 +1,16 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+import react.ComponentClass
+
+external interface BrowserRouterProps : RouterProps {
+    var basename: String
+    var forceRefresh: Boolean
+    var getUserConfirmation: GetUserConfirmation?
+    var keyLength: Int
+}
+
+external val BrowserRouter: ComponentClass<BrowserRouterProps>
+
