+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/History.kt
@@ -1,0 +1,17 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+external interface History {
+    val length: Int
+    val action: String
+    var location: Location
+
+    fun push(path: String, state: dynamic = definedExternally)
+    fun replace(path: String, state: dynamic = definedExternally)
+    fun go(n: Int)
+    fun goBack()
+    fun goForward()
+}
+
