+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/Link.kt
@@ -1,0 +1,16 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+import react.ComponentClass
+import react.PropsWithChildren
+
+external interface LinkProps : PropsWithChildren {
+    var to: String
+    var replace: Boolean
+    var className: String?
+}
+
+external val Link: ComponentClass<LinkProps>
+
