+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/Location.kt
@@ -1,0 +1,10 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+external interface Location {
+    var hash: String
+    var pathname: String
+    var search: String
+}
