+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/Match.kt
@@ -1,0 +1,11 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+external interface Match {
+    var isExact: Boolean
+    var url: String
+    var path: String
+    var params: Params
+}
