+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/NavLink.kt
@@ -1,0 +1,17 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+import react.ComponentClass
+
+external interface NavLinkProps : LinkProps {
+    var activeClassName: String
+    var exact: Boolean
+    var strict: Boolean
+    var isActive: ((match: Match?, location: Location) -> Boolean)?
+    var location: Location
+}
+
+external val NavLink: ComponentClass<NavLinkProps>
+
