+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/Redirect.kt
@@ -1,0 +1,17 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+import react.ComponentClass
+import react.Props
+
+external interface RedirectProps : Props {
+    var from: String?
+    var to: String
+    var push: Boolean
+    var exact: Boolean
+    var strict: Boolean
+}
+
+external val Redirect: ComponentClass<RedirectProps>
