+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/RouterProps.kt
@@ -1,0 +1,12 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+import react.PropsWithChildren
+
+external interface RouterProps : PropsWithChildren {
+    var history: History
+}
+
+
