+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/Switch.kt
@@ -1,0 +1,10 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+import react.ComponentClass
+import react.PropsWithChildren
+
+external val Switch: ComponentClass<PropsWithChildren>
+
